Refactor code into helpers
